Change header, remove container padding and adjust font size
The task is broad aimed at frontend issues with the new Natmus site. This commit contains a new header with logo and logotype. We also add svg menu icon with "open" color.

To use more space on mobile, we remove container class padding. This may have implications on later use, but it's the best way to achieve the wanted look.

Font base size has been decreased. Some headings also need a .hx class to be even smaller.

Tables have been styled in a run-in style. This is not completed yet, because some fields demand special handling.

LP#15202477 and LP#15771745