Adjust header to desktop view
There is a new desktop mockup, which requires some changes in markup and classes (cf. LP#15771745).

The tricky part is to change rendering order of logos area and search/menu area. We use Bootstrap rows and cols to do it. But we have to set menu parent rows to display static - otherwise we can't overlay the screen "on menu open". If this creates cross browser/device problems, we may use another element for overlay.

LP#15771745 and LP#15202477