Going in another direction.
Instead of having two aggregations (single and multifield) that has to be merged
in a post-process, we combine the two aggregations into a single aggreagation
which will allow us to skip the post-process as we get the correct counts right
away.

This and the subsequent commits cleans up after a quick poc version of the
aggregation.

This commit removes the postprocess functionallity and fully commits to a merged
approach.

KB-498