Fix mixed content images
Using image.url for CloudinaryImages apparently uses the raw HTTP url even when cloudinary has been configured to be secure.
The `url` can be seen as an internal reference to the image, rather than what should be shown. In order to go via the "secure"
layer, we need to call _.image.src() instead of .image.url. This ensures that images are HTTPS.

I did a search for all occurrances of .url, found all that referred to images, and replaced the URL references.