Add og:image:secure_url tag
There is an open graph image (shows up when page is shared on e.g. Facebook), and this adds a HTTPS-url for these as well, so sites trying to index this page on the open graph can use the secure version. It is recommended practice to also keep a pure HTTP URL available.