Add that map slant! when it is after a jumbo
On the frontpage, when a map view follows a jumbo, it should be taller and have a slant at the top, that followes the slant of the bottom of the jumbo.

In order to achieve this, we had to rewrite the jumbo slant so it is always a set number of degrees (-5.5deg to be precise).

Also mobile friendly: more distance between jumbo and map on smaller screens; no slant on smallest screens.