Fix edge cases in search input
The search queries sent to Elastic Search are interpreted as Lucene syntax, which means that some characters will result in undefined behavior. For example, + must be between words, - removes terms, etc etc.
In this commit we start stripping these values so they dont result in unexpected behavior (users cannot be expected to know or understand (or guess that we use) Lucene).
Specifically, we replace them with spaces, so we dont end up
concatenating words that should not be concatenated. E.g. magasin+okay
will result in magasin okay rather than magasinokay.