Fix issue of mounting windows dir into linux container
This has been a development issue for a while: when working on windows, setting volume to be the local file system causes some issues, in particular with new changes to npm where checksums in package-lock files are no the same on windows vs linux, and when node-gyp builds result in different bindings on windows or linux.
This fix introduces a new script, install-then-start, which runs install and then start. By using this as the command (instead of plain start) when the docker container runs, the following happens: (1) the container starts; (2) the filesystem is mounted over, so any previously installed node_modules are now replaced with whatever is on the windows filesystem; (3) the install script runs, which replaces the node_modules with proper linux (can-be-used-in-container) node_moduels; (4) start runs in the container, and can indeed use the node_modules as they were built/checksummed on linux.