Remove use of types, use indices instead
The support for the `type` keyword has been removed in elasticsearch. The recommended approach is to use separate indices for separate types of data, so that is what we are now doing: we split the es.index into two incdices, es.assetIndex and es.seriesIndex.

This commit replaces all reference to es.index by making appropriate changes.