use an exact (term) query to look up series containing an asset
The old lookup method was there for historical reasons because the values at least in beta werent keyword values -- but seeing as we are completely reindexing everything, we now ensure the assets field contains proper keywords, so we can do exact lookups instead.

This all happened because we had some code in beta for a while treated
the assets field wrong, and elasticsearch keeps data types forever
(until index destroyed) once set.